[Umstrukturierung]
The folder `input` in `Linux` will be renamed to `src` and moved to the
root directory of the repo.

The files `Setlist.tex` and `setlist.cls` will also be renamed to
`Chordbook.tex` and `chordbook.cls` - all crossreferences between the
main tex file and the sources will be updated, too.