[Folder structure update]
The Linux Folder is emptied except for the Makefile.

Compiling works the same way, run `make` after `cd`-ing into Linux/ and
the pdf file should be compiled right away.

The Tex-File was edited in the way of changing the path to the .cls file
and also the Makefile was edited according to a changed path to the .tex
file, the main .tex file of course.