[ revamp ] this is complete revamp of the document structure
the source folder `src` will be reorganized in the way, that for each
song there will be a folder, which contains two subfolders (`LaTeX` and
`lilypond`). the LaTeX folder will contain the files, former placed in
`src`, but for each song. in my opinion this is a cleaner way to set up
this document

the lilypond folder will always contain a gitignore will, containing PDF
and MIDI. this will prevent midi and pdf files of the lilypond compiler
going into the repository

beside that, for the song 'gods gonna cut you down' there is a lilypond
file added + a makefile for compiling it and managing its output

the 'addToSetlist.sh' is to be revamped as well, for now it is not made
executable anymore ... as some kind of hint.