[ Unix Makefile ] adding publish section
this section does some comfort-stuff for myself. it is not meant to be
for the public :)

the makefile, if started with the option `public`, will create a folder
`PUBLISH` in the repo's root directory and copy the (normal|EXTENDEDD)
chordbook and the rehearsal MIDI files into it.

at last, it will create a (tar-ball|zip-file) from the `PUBLISH` folder
and move it to the root directory as well