[ songs package workaround ] problem with hyperref
According to the maintainer of the songs package (Kevin Hamlen), the
hyperref package a) needs to be loaded and b) before loading the songs
package.

This is even weird, because the error message / the part of the songs
package mentioned in the error message, has nothing to do with hyperref.

This commit is a dirty hack, though, because it will put all the
songs-specific parts into the main document and off the class file.

The good news is, the songs.sty in the `Unix` folder will be gone.