[ Extended Chordbook ] disabling (almost) every song ...
... except for:

- Stille Nacht (Larp)
- Der Wandersmann
- Das Bier ich in der Rechten trug

I will enable songs, when they are ready or in testing.

Beside, the width of the included pdfs are adjustet to 85% instead of
90%.
The to be included pdfs will get a subsection and each pdf, which is
longer than one page needs to be included twice:

1.) with only the first page and *with* the subsection
2.) page 2 till the end *without* the subsection

So, in most cases there will be more than two pages, I hope, but in case
there are not, you need to tweak that manually. At least for now.